+++ darts/tests/explainability/test_shap_explainer.py
@@ -13,7 +13,10 @@ from sklearn.preprocessing import MinMaxScaler
 from darts import TimeSeries
 from darts.dataprocessing.transformers import Scaler
 from darts.explainability.explainability_result import ShapExplainabilityResult
-from darts.explainability.shap_explainer import MIN_BACKGROUND_SAMPLE, SKLearnExplainer
+from darts.explainability.sklearn_explainer import (
+    MIN_BACKGROUND_SAMPLE,
+    SKLearnExplainer,
+)
 from darts.models import (
     CatBoostModel,
     ExponentialSmoothing,
